renamed to boundary departure checker
