+++ planning/autoware_boundary_departure_checker/CMakeLists.txt
@@ -1,11 +1,11 @@
 cmake_minimum_required(VERSION 3.14)
-project(autoware_departure_checker)
+project(autoware_boundary_departure_checker)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
 
 ament_auto_add_library(${PROJECT_NAME} SHARED
-  src/departure_checker.cpp
+  src/boundary_departure_checker.cpp
   src/utils.cpp
 )
