+++ planning/autoware_boundary_departure_checker/include/autoware/boundary_departure_checker/boundary_departure_checker.hpp
@@ -15,7 +15,7 @@
 #ifndef AUTOWARE__DEPARTURE_CHECKER__DEPARTURE_CHECKER_HPP_
 #define AUTOWARE__DEPARTURE_CHECKER__DEPARTURE_CHECKER_HPP_
 
-#include "autoware/departure_checker/parameters.hpp"
+#include "autoware/boundary_departure_checker/parameters.hpp"
 
 #include <autoware_utils/system/time_keeper.hpp>
 #include <autoware_vehicle_info_utils/vehicle_info_utils.hpp>
