+++ planning/autoware_boundary_departure_checker/include/autoware/boundary_departure_checker/utils.hpp
