+++ planning/autoware_boundary_departure_checker/test/test_create_vehicle_passing_areas.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/departure_checker/utils.hpp"
+#include "autoware/boundary_departure_checker/utils.hpp"
 
 #include <Eigen/Core>
 #include <autoware_utils/geometry/boost_geometry.hpp>
