+++ planning/behavior_path_planner/autoware_behavior_path_planner_common/package.xml
@@ -44,7 +44,7 @@
   <buildtool_depend>autoware_cmake</buildtool_depend>
 
   <depend>autoware_adapi_v1_msgs</depend>
-  <depend>autoware_departure_checker</depend>
+  <depend>autoware_boundary_departure_checker</depend>
   <depend>autoware_freespace_planning_algorithms</depend>
   <depend>autoware_internal_planning_msgs</depend>
   <depend>autoware_interpolation</depend>
