+++ planning/behavior_path_planner/autoware_behavior_path_start_planner_module/include/autoware/behavior_path_start_planner_module/shift_pull_out.hpp
@@ -19,7 +19,7 @@
 #include "autoware/behavior_path_start_planner_module/pull_out_planner_base.hpp"
 #include "autoware_utils/system/time_keeper.hpp"
 
-#include <autoware/departure_checker/departure_checker.hpp>
+#include <autoware/boundary_departure_checker/boundary_departure_checker.hpp>
 
 #include <autoware_internal_planning_msgs/msg/path_with_lane_id.hpp>
